refactored methods and processes
