+++ mindsdb/integrations/handlers/lightwood_handler/lightwood_handler/__about__.py
@@ -1,6 +1,6 @@
 __title__ = 'MindsDB Lightwood handler'
 __package_name__ = 'mindsdb_lightwood_handler'
-__version__ = '0.0.1'
+__version__ = '1.0.0'
 __description__ = "MindsDB handler for Lightwood"
 __author__ = 'MindsDB Inc'
 __github__ = 'https://github.com/mindsdb/mindsdb'
