+++ mindsdb/integrations/handlers/lightwood_handler/lightwood_handler/utils.py
@@ -1,5 +1,6 @@
 import gc
 import sys
+import json
 import dill
 
 from lightwood.api.high_level import _module_from_code
@@ -21,7 +22,6 @@ def unpack_jsonai_old_args(json_ai_override):
 
 def load_predictor(predictor_dict, name):
     try:
-        module_name = None
         return dill.loads(predictor_dict['predictor'])
     except Exception as e:
         module_name = str(e).lstrip("No module named '").split("'")[0]
@@ -34,3 +34,46 @@ def load_predictor(predictor_dict, name):
         gc.collect()
         _module_from_code(predictor_dict['code'], module_name)
         return dill.loads(predictor_dict['predictor'])
+
+
+def rep_recur(org: dict, ovr: dict):
+    for k in ovr:
+        if k in org:
+            if isinstance(org[k], dict) and isinstance(ovr[k], dict):
+                rep_recur(org[k], ovr[k])
+            else:
+                org[k] = ovr[k]
+        else:
+            org[k] = ovr[k]
+
+
+def brack_to_mod(ovr):
+    if not isinstance(ovr, dict):
+        if isinstance(ovr, list):
+            for i in range(len(ovr)):
+                ovr[i] = brack_to_mod(ovr[i])
+        elif isinstance(ovr, str):
+            if '(' in ovr and ')' in ovr:
+                mod = ovr.split('(')[0]
+                args = {}
+                if '()' not in ovr:
+                    for str_pair in ovr.split('(')[1].split(')')[0].split(','):
+                        k = str_pair.split('=')[0].strip(' ')
+                        v = str_pair.split('=')[1].strip(' ')
+                        args[k] = v
+
+                ovr = {
+                    'module': mod,
+                    'args': args
+                }
+            elif '{' in ovr and '}' in ovr:
+                try:
+                    ovr = json.loads(ovr)
+                except Exception:
+                    pass
+        return ovr
+    else:
+        for k in ovr.keys():
+            ovr[k] = brack_to_mod(ovr[k])
+
+    return ovr
